+++ src/Module.php
@@ -1,0 +1,42 @@
+<?php
+/**
+ * @link      http://github.com/zendframework/zend-serializer for the canonical source repository
+ * @copyright Copyright (c) 2005-2016 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   http://framework.zend.com/license/new-bsd New BSD License
+ */
+
+namespace Zend\Serializer;
+
+class Module
+{
+    /**
+     * Return default zend-serializer configuration for zend-mvc applications.
+     */
+    public function getConfig()
+    {
+        $provider = new ConfigProvider();
+
+        return [
+            'service_manager' => $provider->getDependencyConfig(),
+        ];
+    }
+
+    /**
+     * Register a specification for the SerializerAdapterManager with the ServiceListener.
+     *
+     * @param \Zend\ModuleManager\ModuleEvent
+     * @return void
+     */
+    public function init($event)
+    {
+        $container = $event->getParam('ServiceManager');
+        $serviceListener = $container->get('ServiceListener');
+
+        $serviceListener->addServiceManager(
+            'SerializerAdapterManager',
+            'serializers',
+            'Zend\ModuleManager\Feature\SerializerProviderInterface',
+            'getSerializerConfig'
+        );
+    }
+}